Revert "Deployments page"
