+++ src/main/resources/assets/js/components/TabPaneComponent.jsx
@@ -6,8 +6,6 @@ define([
   "use strict";
 
   return React.createClass({
-    name: "TabPaneComponent",
-
     propTypes: {
       isActive: React.PropTypes.bool,
       onActivate: React.PropTypes.func
