+++ src/main/resources/assets/js/components/TogglableTabsComponent.jsx
@@ -2,42 +2,38 @@
 
 define([
   "React",
+  "Underscore",
   "jsx!components/NavTabsComponent"
-], function(React, NavTabsComponent) {
+], function(React, _, NavTabsComponent) {
   "use strict";
 
   return React.createClass({
-    name: "TogglableTabsComponent",
-
-    propTypes: {
-      activeTabId: React.PropTypes.string.isRequired,
-      className: React.PropTypes.string,
-      onTabClick: React.PropTypes.func,
-      tabs: React.PropTypes.array
+    getInitialState: function() {
+      return {
+        activeTabId: _.isArray(this.props.children) ?
+          this.props.children[0].props.id :
+          this.props.children.props.id
+      };
     },
-
+    onTabClick: function(id) {
+      this.setState({
+        activeTabId: id
+      });
+    },
     render: function() {
       var childNodes = React.Children.map(this.props.children, function(child) {
         return React.addons.cloneWithProps(child, {
-          isActive: (child.props.id === this.props.activeTabId)
+          isActive: (child.props.id === this.state.activeTabId)
         });
       }, this);
-
-      var nav;
-      if (this.props.onTabClick != null && this.props.tabs != null) {
-        /* jshint trailing:false, quotmark:false, newcap:false */
-        nav = (
-          <NavTabsComponent
-            activeTabId={this.props.activeTabId}
-            onTabClick={this.props.onTabClick}
-            tabs={this.props.tabs} />
-        );
-      }
 
       /* jshint trailing:false, quotmark:false, newcap:false */
       return (
         <div className={this.props.className}>
-          {nav}
+          <NavTabsComponent
+            activeTabId={this.state.activeTabId}
+            onTabClick={this.onTabClick}
+            tabs={this.props.tabs} />
           <div className="tab-content">
             {childNodes}
           </div>
