+++ src/main/resources/assets/js/models/AppCollection.js
@@ -6,7 +6,6 @@
 
   return SortableCollection.extend({
     model: App,
-
     initialize: function(models, options) {
       this.options = options;
       this.setComparator("-id");
