+++ src/main/resources/assets/js/models/Deployment.js
@@ -1,34 +1,14 @@
 define([
-  "Backbone",
-  "Underscore"
-], function(Backbone, _) {
+  "Backbone"
+], function(Backbone) {
   "use strict";
 
   return Backbone.Model.extend({
     defaults: function() {
       return {
-        currentStep: 0,
-        currentActions: [],
-        affectedApps: [],
-        totalSteps: 0,
+        affectedApplications: [],
         steps: []
       };
-    },
-
-    currentActionsString: function() {
-      var actions = this.get("currentActions");
-      var actionNames = actions.map(function(action) {
-        return action.action;
-      });
-      return actionNames.join(", ");
-    },
-
-    affectedAppsString: function() {
-      var actions = this.get("currentActions");
-      var apps = actions.map(function(action) {
-        return action.apps;
-      });
-      return apps.join(", ");
     },
 
     initialize: function(options) {
